requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
numpy
streamlit
pytest
python-dotenv
langchain
langchain-openai
langchain-chroma
langchain-community
